money

+++ python_introduction/finance_calculator.py
@@ -1,5 +1,5 @@
-Monthly_income = int(input("Enter your monthly income: ")) 
-Monthly_expences = int(input("Enter our total expences: "))
+Monthly_income = float(input("Enter your monthly income: ")) 
+Monthly_expences = float(input("Enter our total expences: "))
 Monthly_savings = Monthly_income - Monthly_expences
 Projected_Savings = Monthly_savings * 12 + (Monthly_savings * 12 * 0.05)
 print("Your monthly savings are $",Monthly_savings)
